pyproject.toml
[project]
name = "pipeedge"
version = "0.1.0"
description = "PipeEdge"
authors = [
    { "name" = "Yang Hu", "email" = "yhu210003@usc.edu" },
    { "name" = "Connor Imes", "email" = "cimes@isi.edu" },
    { "name" = "Haonan Wang", "email" = "haonanwa@usc.edu" },
]
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "apphb>=0.1.0",
    "energymon>=0.1.0",
    "numpy",
    "Pillow",
    "psutil",
    "PyYAML",
    "requests",
    "torch>=1.8.0",
    "transformers>=4.6.0",
]

[project.optional-dependencies]
# Required by the project source "baseline" directory
baseline = [
    "timm>=0.3.2",
]
# Required by the project source "tools" directory
tools = [
    "timm>=0.3.2",
]

[build-system]
# Work around current py-build-cmake issues (as of 0.0.10):
# Support Python 3.7: https://github.com/tttapa/py-build-cmake/issues/5
# Enforce min cmake version for CLI compatibility: https://github.com/tttapa/py-build-cmake/issues/6
requires = ["cmake>=3.15", "py-build-cmake @ git+https://github.com/cimes-isi/py-build-cmake.git@py3.7"]
build-backend = "py_build_cmake.build"

[tool.py-build-cmake.module]
directory = "src"

[tool.py-build-cmake.sdist]
include = [
    "partition/CMakeLists.txt",
    "partition/sched-pipeline.cpp",
    "partition/yaml_types.h",
    "partition/README.md",
]

[tool.py-build-cmake.cmake]
minimum_version = "3.8"
build_type = "RelWithDebInfo"
source_path = "partition"
build_args = ["-j"]
install_components = ["PipeEdge_Runtime"]
